refactor(algorithms): keep DRPO module path
Restore ARDRPO to the existing drpo module path while keeping the FlowGRPO rename.
