+++ unirl/algorithms/__init__.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 from .ar_grpo import ARGRPO
 from .base import AlgorithmStepResult, StageAlgorithm
 from .flowgrpo import FlowGRPO
-from .ar_drpo import ARDRPO
+from .drpo import ARDRPO
 from .flowdppo import FlowDPPO, FlowDPPOConfig
 from .nft import DiffusionNFT, DiffusionNFTConfig
